requirements.txt
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
pdfplumber>=0.9.0
openai>=0.27.0

fastapi>=0.95.0
uvicorn[standard]>=0.22.0

scikit-learn>=1.2.2

pandas>=2.0.1
